+++ WinnerWinner_Final/tilemap.py
@@ -55,14 +55,16 @@ class Camera:
         else:
             return rect.move(self.camera.topleft)
 
-
     def update(self, target):
         x = -target.rect.centerx + int(WIDTH / 2)
         y = -target.rect.centery + int(HEIGHT / 2)
-
-        # limit scrolling to map size
+        # fix scrolling to map size
         x = min(0, x)  # left
         y = min(0, y)  # top
         x = max(-(self.width - WIDTH), x)  # right
         y = max(-(self.height - HEIGHT), y)  # bottom
         self.camera = pg.Rect(x, y, self.width, self.height)
+
+    # def apply_rect_minimap_camera(self, rect, mm_width, mm_height, mm_center):
+    #     minimap_camera = pg.Rect(mm_center.x, mm_center.y, mm_width, mm_height)
+    #     return rect.move(minimap_camera.topleft)